Fixed a bug where a revision spanning many mapping entries had a chance to only update one of them.
The problem was that if you have two mapping entries with the following
source paths:

/trunk/a
/trunk/a-b

Then /trunk/a would also match commits to /trunk/a-b because it matches
the start of the changed path. The solution was to reverse the order it
checks the source paths, so the most specific ones always are checked
first.